Break out keydown inserter
